py ver: no deletions from queue
How it works:
Deletions from the queue were actually preparation for lowering the priority.
It means that subsequent openq/openset add is storing the same
node with lower priority. Such pair (lower_pri, node) is then
picked up by another pop() later. At the same time, node is
deleted from openset, so if we decide to omit openq.delete_by_value(),
then we just have 'hanging' pairs in the queue in the form
(old_pri, node_deleted_from_openset),
which just can be detected during pop().

Such strategy simplifies the code (at the cost of complicating pop())
and gives a minimal speedup.

To clean up the code, the structures openset & openq are blended into
one, with __len__, __contains__ - internal set is now
the primary source of information whether the item is in the structure
or not.